update : collect are now working properly
update : errors messages on modal to not let the volunteer have null inputs

update : filters options on manage-users Page ( by city and search option )

add : add Modal to create new volunteers ( in work )

need : some fixes on the edit and add modal, they're not working properly

+++ Projet_Adaction_Front/src/app/manage-users/manage-users.ts
@@ -3,15 +3,26 @@ import { MatDialog } from '@angular/material/dialog';
 import { CommonModule } from '@angular/common';
 import { ManageUsersEditModal } from '../manage-users-edit-modal/manage-users-edit-modal';
 import { Volunteer, VolunteerService } from '../services/volunteer';
+import { MatIconModule } from '@angular/material/icon'
+import { MatButtonModule } from '@angular/material/button';
+import { ManageUsersAddModal } from '../manage-users-add-modal/manage-users-add-modal';
+import { FormsModule } from '@angular/forms';
+import { MatSelectModule } from '@angular/material/select'
+import { MatFormFieldModule } from '@angular/material/form-field';
+import { MatInputModule } from '@angular/material/input';
 
 @Component({
+  standalone: true,
   selector: 'app-manage-users',
-  imports: [CommonModule],
+  imports: [CommonModule, MatIconModule, MatButtonModule, FormsModule, MatSelectModule, MatFormFieldModule, MatInputModule],
   templateUrl: './manage-users.html',
   styleUrls: ['./manage-users.css']
 })
 export class ManageUsers implements OnInit {
   volunteers : Volunteer[] = [];
+  searchQuery: string = '';
+  selectedCity: string = '';
+  cities: string[] = [];
 
   constructor(private volunteerService : VolunteerService, private dialog: MatDialog) {}
 
@@ -23,6 +34,7 @@ export class ManageUsers implements OnInit {
     this.volunteerService.getAllVolunteers().subscribe(
       (volunteers) =>  {
         this.volunteers = volunteers; 
+        this.extractCities();
       },
       (error) => {
         console.error("Erreur lors du chargement des bénévoles :", error);
@@ -31,15 +43,31 @@ export class ManageUsers implements OnInit {
   }
 
   deleteVolunteer(id: number): void {
-    this.volunteerService.deleteVolunteer(id).subscribe(
-      () => {
-        this.loadVolunteers();
-      },
-      (error) => {
-        console.error("Erreur lors de la suppression :", error);
-      }
-    );
+    if(confirm("Êtes-vous sûr de vouloir supprimer ce bénévole ?")) {
+      this.volunteerService.deleteVolunteer(id).subscribe(
+        () => {
+          this.loadVolunteers();
+        },
+        (error) => {
+          console.error("Erreur lors de la suppression :", error);
+        }
+      );
+    }
   }
+
+  extractCities(): void {
+    this.cities = [...new Set(this.volunteers.map(volunteer => volunteer.location))];
+  }
+
+  get filteredVolunteers(): Volunteer[] {
+    return this.volunteers.filter(volunteer => {
+    const matchesSearch = volunteer.firstname.toLowerCase().includes(this.searchQuery.toLowerCase()) || 
+                          volunteer.lastname.toLowerCase().includes(this.searchQuery.toLowerCase());
+    
+    const matchesCity = this.selectedCity ? volunteer.location === this.selectedCity : true;
+    return matchesSearch && matchesCity;
+    })
+  };
 
   openEditModal(volunteer: Volunteer): void {
     const dialogRef = this.dialog.open(ManageUsersEditModal, {
@@ -59,5 +87,23 @@ export class ManageUsers implements OnInit {
         );
       }
     });
-  } 
+  }
+
+  openAddModal(): void {
+    const dialogRef = this.dialog.open(ManageUsersAddModal, {
+      width: '400px',
+    });
+
+    dialogRef.afterClosed().subscribe((newVolunteer) => {
+      if(newVolunteer) {
+        this.volunteerService.addVolunteer(newVolunteer).subscribe(
+          () => {
+            this.loadVolunteers();
+          }, (error) => {
+            console.error("Erreur lors de l'ajout :" + error)
+          }
+        );
+      }
+    });
+  }
 }
